tests: Move `test_match_debug` to test_stbt_debug.py
